支持多颜色附件 + 抽象 GL 工具函数 + 窗口/日志/Ref 整理
- OpenGLFramebuffer: 重写为支持多颜色附件 (m_ColorAttachments/m_ColorAttachmentFormats)，抽象 Utils 工具函数 (CreateTextures/AttachColorTexture/AttachDepthTexture 等)
- FramebufferSpecification 支持 Attachments 列表，不再使用单一 Format 字段
- BindTexture 增加 attachmentIndex 参数指定附件索引
- 析构/Resize lambda 改用 Ref 持活对象
- 新增 m_Width/m_Height 成员，Resize 时直接比较减少 spec 依赖
- WindowsWindow: Maximize/VSync 延迟到 Application 初始化完成后调用
- Application::OnWindowResize: 跳过 NoResize 标记的 Framebuffer
- RefCounted: m_RefCount 改为 atomic<uint32_t>，支持跨线程安全
- Ref 调试: 添加互斥锁保护 LiveReferences，Shutdown 时跳过断言
- Log.h: glm::format_as 移至 glm 命名空间内
- EditorCamera: 成员统一初始化默认值
- PhysicsSettingsWindow: 使用 Prism/ImGui/ImGui.h
- Material.cs: 移除 DefaultMaterial 属性

+++ Prism/src/Prism/Renderer/Renderer2D.cpp
@@ -191,7 +191,7 @@ namespace Prism
             s_Data.Stats.DrawCalls++;
         }
 
-        dataSize = (uint8_t*)s_Data.LineVertexBufferPtr - (uint8_t*)s_Data.LineVertexBufferBase;
+        dataSize = (uint32_t)((uint8_t*)s_Data.LineVertexBufferPtr - (uint8_t*)s_Data.LineVertexBufferBase);
         if (dataSize)
         {
             s_Data.LineVertexBuffer->SetData(s_Data.LineVertexBufferBase, dataSize);
